accessibility_backend/requirements.txt
fastapi[all]
python-dotenv
openai
pydantic